ingestion/requirements.txt
requests
google-cloud-bigquery
python-dotenv
dbt-bigquery
